Struct chunk_store::ChunkStore

pub struct ChunkStore {
    // some fields omitted
}

ChunkStore is a collection for holding all data chunks. Implements a maximum disk usage to restrict storage.

The data chunks are deleted when the ChunkStore goes out of scope.

Methods

impl ChunkStore

fn new_in<P: AsRef<Path>>(root: P, prefix: &str, max_space: usize) -> Result<ChunkStore, Error>

Creates new ChunkStore with max_space allowed storage space.

The data are stored in a temporary directory that contains prefix in its name and is placed in the root directory. If root doesn't exist, it will be created.

fn new(prefix: &str, max_disk_usage: usize) -> Result<ChunkStore, Error>

Creates new ChunkStore with max_space allowed storage space.

The data are stored in a temporary directory that contains prefix in its name and is placed in the system temp directory.

fn put(&mut self, name: &XorName, value: &[u8]) -> Result<(), Error>

Stores a new data chunk under name.

If there is not enough storage space available, returns Error::NotEnoughSpace. In case of an IO error, it returns Error::Io.

If the name already exists, it will be overwritten.

fn delete(&mut self, name: &XorName) -> Result<(), Error>

Deletes the data chunk stored under name.

If the name doesn't exist, it does nothing and returns Ok. In case of an IO error it returns Error::Io.

fn get(&self, name: &XorName) -> Result<Vec<u8>, Error>

Reads a data chunk stored under name.

If the name doesn't exist, returns Error::ChunkNotFound. In Case of an IO error, returns Error::Io.

fn has_chunk(&self, name: &XorName) -> bool

Tests if a data chunk with name is stored in this ChunkStore.

fn names(&self) -> Vec<XorName>

Lists names of all data chunks currently stored in this ChunkStore.

fn max_space(&self) -> usize

Returns the maximum amount of storage space available for this ChunkStore.

fn used_space(&self) -> usize

Returns the amount of storage space already used by this ChunkStore.

fn has_space(&self, required_space: usize) -> bool

Tests if there is enough storage space to store a data chunk of required_space bytes.